// src/enums/token-type.enum.ts
export enum TokenType {
    VALUE = 'VALUE',

    LEFT_PAREN = 'LEFT_PAREN',
    RIGHT_PAREN = 'RIGHT_PAREN',

    LEFT_BRACE = 'LEFT_BRACE',
    RIGHT_BRACE = 'RIGHT_BRACE',

    LEFT_BRACKET = 'LEFT_BRACKET',
    RIGHT_BRACKET = 'RIGHT_BRACKET',

    STAR = 'STAR',
    QUESTION = 'QUESTION',
    PLUS = 'PLUS',
    COMMA = 'COMMA',

    PIPE = 'PIPE',
    CARET = 'CARET',
}